requirements.txt
sentence-transformers
faiss-cpu
flask
openai
bs4